requirements.txt
discord-webhook
psycopg2-binary
discord.py
json5